requirements.txt
aniso8601==8.1.1
asgiref==3.4.1
certifi==2021.5.30
chardet==4.0.0
Django==3.2.5
django-cors-headers==3.7.0
django-graphql-jwt==0.3.2
django-jet-reboot==1.1.1
django-sslserver==0.22
djangorestframework==3.12.4
et-xmlfile==1.1.0
feedparser==6.0.8
google-api-python-client==1.4.1
graphene==3.0b7
graphene-django==3.0.0b7
graphene-file-upload==1.3.0
graphql-core==3.1.5
graphql-relay==3.1.0
gunicorn==20.1.0
httplib2==0.19.1
idna==2.10
mysqlclient==2.0.3
oauth2client==4.1.3
openpyxl==3.0.7
promise==2.3
psycopg2-binary==2.9.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyJWT==2.1.0
pyparsing==2.4.7
pytz==2020.4
razorpay==1.2.0
requests==2.25.1
rsa==4.7.2
Rx==1.6.1
sgmllib3k==1.0.0
shortid==0.1.2
singledispatch==3.4.0.3
six==1.15.0
sqlparse==0.4.1
supervisor==4.2.2
text-unidecode==1.3
Unidecode==1.1.1
uritemplate==3.0.1
urllib3==1.26.6
waitress==2.0.0